New design
So I went to HackRU and ended up needing a Client/Server socket utility.
This commit is essentially the result of a few hours of work I could
afford to spend on such utility. I edited out the majority of temporary
code, but this still needs a lot of work before it's useable.